+++ TinkoffASDKUI/TinkoffASDKUI/Colors/UIColor+Dynamic.swift
@@ -32,7 +32,8 @@ extension UIColor {
     static func dynamicColor(dynamic: Dynamic) -> UIColor {
         if #available(iOS 13.0, *) {
             return UIColor(dynamicProvider: {
-                $0.userInterfaceStyle == .dark ? dynamic.dark : dynamic.light
+                postUserInterfaceStyleDidChangeIfNeeded(style: $0.userInterfaceStyle)
+                return $0.userInterfaceStyle == .dark ? dynamic.dark : dynamic.light
             })
         }
         return dynamic.light
@@ -45,5 +46,15 @@ extension UIColor {
             })
         }
         return dynamic.dark
+    }
+
+    private static func postUserInterfaceStyleDidChangeIfNeeded(style: UIUserInterfaceStyle) {
+        guard UIScreen.main.traitCollection.userInterfaceStyle != style
+        else { return }
+        NotificationCenter.default.post(
+            name: .userInterfaceStyleDidChange,
+            object: self,
+            userInfo: [Notification.Keys.value: style]
+        )
     }
 }