+++ TinkoffASDKUI/TinkoffASDKUI/UIComponents/CardField/CardFieldView + Config.swift
@@ -66,7 +66,6 @@ extension CardFieldView.Config {
             .set(textColor: ASDKColors.Text.secondary.color)
 
         return Self(
-            data: data.cardFieldData,
             style: .regular,
             dynamicCardIcon: DynamicIconCardView.Model(data: data.dynamicCardIconData),
             expirationTextFieldConfig: TextField.Configuration(
@@ -74,7 +73,7 @@ extension CardFieldView.Config {
                     delegate: data.expirationTextFieldData.delegate,
                     text: data.expirationTextFieldData.text,
                     placeholder: data.expirationTextFieldData.placeholder,
-                    hasClearButton: false,
+                    hasClearButton: true,
                     keyboardType: .decimalPad
                 ),
                 headerLabel: UILabel.Configuration(
@@ -98,7 +97,7 @@ extension CardFieldView.Config {
                     delegate: data.cvcTextFieldData.delegate,
                     text: data.cvcTextFieldData.text,
                     placeholder: data.cvcTextFieldData.placeholder,
-                    hasClearButton: false,
+                    hasClearButton: true,
                     keyboardType: .decimalPad,
                     isSecure: true
                 ),